backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
redis
flask-jwt-extended
werkzeug
flask-cors
pytest
